refactor. 전체적인 래픽토링 1. 네이밍 리팩토링 2. 클래스,메서드별로 분할
