+++ build.gradle
@@ -24,6 +24,7 @@ repositories {
 }
 
 dependencies {
+    // Web
     implementation 'org.springframework.boot:spring-boot-starter-web'
 
     // JPA
