+++ src/main/java/com/example/naejeonhajab/annotation/TrackingTime.java
@@ -1,0 +1,11 @@
+package com.example.naejeonhajab.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface TrackingTime {
+}
