+++ src/main/java/com/example/naejeonhajab/domain/game/lol/dto/res/LolPlayerHistoryResponseSimpleDto.java
@@ -1,6 +1,6 @@
 package com.example.naejeonhajab.domain.game.lol.dto.res;
 
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayerHistory;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayerHistory;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
