+++ src/main/java/com/example/naejeonhajab/domain/game/lol/entity/resultHistory/LolResultLines.java
@@ -1,4 +1,4 @@
-package com.example.naejeonhajab.domain.game.lol.entity.result;
+package com.example.naejeonhajab.domain.game.lol.entity.resultHistory;
 
 import com.example.naejeonhajab.domain.game.lol.dto.common.LolLinesDto;
 import com.example.naejeonhajab.domain.game.lol.dto.common.LolPlayerDto;
