--- src/main/java/com/example/naejeonhajab/domain/game/lol/repository/player/LolLinesRepository.java
@@ -1,11 +1,0 @@
-package com.example.naejeonhajab.domain.game.lol.repository.player;
-
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayer;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolLines;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface LolLinesRepository extends JpaRepository<LolLines, Long> {
-    List<LolLines> findAllByPlayer(LolPlayer player);
-}
