+++ src/main/java/com/example/naejeonhajab/domain/game/lol/repository/playerHistory/LolPlayerRepository.java
@@ -1,7 +1,7 @@
-package com.example.naejeonhajab.domain.game.lol.repository.player;
+package com.example.naejeonhajab.domain.game.lol.repository.playerHistory;
 
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayer;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayerHistory;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayer;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayerHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
