+++ src/main/java/com/example/naejeonhajab/domain/game/lol/repository/resultHistory/LolPlayerResultOutcomeRepository.java
@@ -1,6 +1,6 @@
-package com.example.naejeonhajab.domain.game.lol.repository.result;
+package com.example.naejeonhajab.domain.game.lol.repository.resultHistory;
 
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResultOutcome;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResultOutcome;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LolPlayerResultOutcomeRepository extends JpaRepository<LolPlayerResultOutcome, Long> {
