+++ src/main/java/com/example/naejeonhajab/domain/game/lol/service/AbyssServiceImpl.java
@@ -11,17 +11,17 @@ import com.example.naejeonhajab.domain.game.lol.dto.res.LolPlayerResultHistoryRe
 import com.example.naejeonhajab.domain.game.lol.dto.res.LolTeamResponseDto;
 import com.example.naejeonhajab.domain.game.lol.dto.res.LolPlayerHistoryResponseSimpleDto;
 import com.example.naejeonhajab.domain.game.lol.dto.res.LolPlayerResultHistoryResponseSimpleDto;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayer;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayerHistory;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResult;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResultHistory;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResultOutcome;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayer;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayerHistory;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResult;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResultHistory;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResultOutcome;
 import com.example.naejeonhajab.domain.game.lol.enums.LolType;
-import com.example.naejeonhajab.domain.game.lol.repository.player.LolPlayerHistoryRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.player.LolPlayerRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolPlayerResultHistoryRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolPlayerResultOutcomeRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolPlayerResultRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.playerHistory.LolPlayerHistoryRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.playerHistory.LolPlayerRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolPlayerResultHistoryRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolPlayerResultOutcomeRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolPlayerResultRepository;
 import com.example.naejeonhajab.domain.user.entity.User;
 import com.example.naejeonhajab.security.AuthUser;
 import lombok.RequiredArgsConstructor;
