+++ src/main/java/com/example/naejeonhajab/domain/game/lol/service/TftServiceImpl.java
@@ -3,25 +3,24 @@ package com.example.naejeonhajab.domain.game.lol.service;
 import com.example.naejeonhajab.common.exception.BaseException;
 import com.example.naejeonhajab.common.exception.LolException;
 import com.example.naejeonhajab.common.response.enums.BaseApiResponse;
-import com.example.naejeonhajab.domain.game.lol.dto.common.LolLinesDto;
 import com.example.naejeonhajab.domain.game.lol.dto.common.LolPlayerDto;
 import com.example.naejeonhajab.domain.game.lol.dto.req.LolPlayerHistoryRequestDto;
 import com.example.naejeonhajab.domain.game.lol.dto.req.LolPlayerResultHistoryRequestDto;
 import com.example.naejeonhajab.domain.game.lol.dto.res.*;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolLines;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayer;
-import com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayerHistory;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResult;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResultHistory;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResultOutcome;
-import com.example.naejeonhajab.domain.game.lol.entity.result.LolResultLines;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolLines;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayer;
+import com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayerHistory;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResult;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResultHistory;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResultOutcome;
+import com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolResultLines;
 import com.example.naejeonhajab.domain.game.lol.enums.LolType;
-import com.example.naejeonhajab.domain.game.lol.repository.player.LolPlayerHistoryRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.player.LolPlayerRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolPlayerResultHistoryRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolPlayerResultOutcomeRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolPlayerResultRepository;
-import com.example.naejeonhajab.domain.game.lol.repository.result.LolResultLinesRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.playerHistory.LolPlayerHistoryRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.playerHistory.LolPlayerRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolPlayerResultHistoryRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolPlayerResultOutcomeRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolPlayerResultRepository;
+import com.example.naejeonhajab.domain.game.lol.repository.resultHistory.LolResultLinesRepository;
 import com.example.naejeonhajab.domain.user.entity.User;
 import com.example.naejeonhajab.security.AuthUser;
 import lombok.RequiredArgsConstructor;
