+++ src/main/resources/mapper/LolMapper.xml
@@ -4,7 +4,7 @@
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 <mapper namespace="com.example.naejeonhajab.domain.game.lol.mapper.LolMapper">
-    <insert id="insertPlayerHistory" useGeneratedKeys="true" keyProperty="id" parameterType="com.example.naejeonhajab.domain.game.lol.entity.player.LolPlayerHistory">
+    <insert id="insertPlayerHistory" useGeneratedKeys="true" keyProperty="id" parameterType="com.example.naejeonhajab.domain.game.lol.entity.playerHistory.LolPlayerHistory">
         INSERT INTO lol_player_history (user_id, player_history_title, type, created_at)
         VALUES (#{user.id}, #{playerHistoryTitle}, #{type}, NOW());
     </insert>
@@ -25,7 +25,7 @@
         </foreach>
     </insert>
 
-    <insert id="inserPlayerResultHistory" useGeneratedKeys="true" keyProperty="id" parameterType="com.example.naejeonhajab.domain.game.lol.entity.result.LolPlayerResultHistory">
+    <insert id="inserPlayerResultHistory" useGeneratedKeys="true" keyProperty="id" parameterType="com.example.naejeonhajab.domain.game.lol.entity.resultHistory.LolPlayerResultHistory">
         INSERT INTO lol_player_result_history (user_id, player_result_history_title, type, created_at)
         VALUES (#{user.id}, #{playerResultHistoryTitle}, #{type}, NOW());
     </insert>
